Sentinel: bind to source address if we have one
Sentinel breaks in some environments when multiple source IPs
are available or if the source IP is a floaty vip address.

This changes Sentinel to act like Cluster and use the first
configured bind address for listening as our source address
as well.

Note: this commit is temporary and includes updates to hiredis
that are backported into actual hiredis.

This commit should not be merged into Redis-Prime until hiredis
gets a proper version increase.

The additions to hiredis itself are being tracked at
https://github.com/mattsta/hiredis/tree/add-source-bind-support