Fix potential error when generating INFO string
Clang (scan-build) flags this as an error, but it can't
happen under normal usage because the input is verified
in caller function.  Clang doesn't trust the caller
to always pass correct input though, so it gets registered
as an error in the report.

It's better to remove the possibility of a crash, even if
only technically.  It also makes our error report cleaner
since we know this isn't a problem.